refactor to use Self type annotation
