+++ tests/geometries/linearring_test.py
@@ -15,7 +15,6 @@
 # 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301 USA
 
 """Test the geometry classes."""
-from typing import cast
 
 import pygeoif.geometry as geo
 import pytest
@@ -51,39 +50,30 @@ class TestLinearRing(StdLibrary):
 
     def test_from_string(self) -> None:
         """Test the from_string method."""
-        linear_ring = cast(
-            LinearRing,
-            LinearRing.class_from_string(
-                '<kml:LinearRing xmlns:kml="http://www.opengis.net/kml/2.2">'
-                "<kml:coordinates>0.000000,0.000000 1.000000,0.000000 1.0,1.0 "
-                "0.000000,0.000000</kml:coordinates>"
-                "</kml:LinearRing>",
-            ),
+        linear_ring = LinearRing.class_from_string(
+            '<kml:LinearRing xmlns:kml="http://www.opengis.net/kml/2.2">'
+            "<kml:coordinates>0.000000,0.000000 1.000000,0.000000 1.0,1.0 "
+            "0.000000,0.000000</kml:coordinates>"
+            "</kml:LinearRing>",
         )
 
         assert linear_ring.geometry == geo.LinearRing(((0, 0), (1, 0), (1, 1), (0, 0)))
 
     def test_empty_from_string(self) -> None:
         """Test the from_string method with an empty LinearRing."""
-        linear_ring = cast(
-            LinearRing,
-            LinearRing.class_from_string(
-                '<kml:LinearRing xmlns:kml="http://www.opengis.net/kml/2.2">'
-                "<kml:coordinates></kml:coordinates>"
-                "</kml:LinearRing>",
-            ),
+        linear_ring = LinearRing.class_from_string(
+            '<kml:LinearRing xmlns:kml="http://www.opengis.net/kml/2.2">'
+            "<kml:coordinates></kml:coordinates>"
+            "</kml:LinearRing>",
         )
 
         assert linear_ring.geometry.is_empty
 
     def test_no_coordinates_from_string(self) -> None:
         """Test the from_string method with an empty LinearRing."""
-        linear_ring = cast(
-            LinearRing,
-            LinearRing.class_from_string(
-                '<kml:LinearRing xmlns:kml="http://www.opengis.net/kml/2.2">'
-                "</kml:LinearRing>",
-            ),
+        linear_ring = LinearRing.class_from_string(
+            '<kml:LinearRing xmlns:kml="http://www.opengis.net/kml/2.2">'
+            "</kml:LinearRing>",
         )
 
         assert linear_ring.geometry.is_empty
@@ -103,15 +93,12 @@ class TestLinearRing(StdLibrary):
 
     def test_mixed_2d_3d_coordinates_from_string_relaxed(self) -> None:
         """Test the from_string method with mixed 2D and 3D coordinates."""
-        linear_ring = cast(
-            LinearRing,
-            LinearRing.class_from_string(
-                '<kml:LinearRing xmlns:kml="http://www.opengis.net/kml/2.2">'
-                "<kml:coordinates>0.000000,0.000000 1.000000,0.000000 1.0,1.0 "
-                "0.000000,0.000000 1.000000,2.000000,3.000000</kml:coordinates>"
-                "</kml:LinearRing>",
-                strict=False,
-            ),
+        linear_ring = LinearRing.class_from_string(
+            '<kml:LinearRing xmlns:kml="http://www.opengis.net/kml/2.2">'
+            "<kml:coordinates>0.000000,0.000000 1.000000,0.000000 1.0,1.0 "
+            "0.000000,0.000000 1.000000,2.000000,3.000000</kml:coordinates>"
+            "</kml:LinearRing>",
+            strict=False,
         )
 
         assert not linear_ring
