+++ tests/geometries/linestring_test.py
@@ -15,7 +15,6 @@
 # 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301 USA
 
 """Test the geometry classes."""
-from typing import cast
 
 import pygeoif.geometry as geo
 import pytest
@@ -51,17 +50,14 @@ class TestLineString(StdLibrary):
 
     def test_from_string(self) -> None:
         """Test the from_string method."""
-        linestring = cast(
-            LineString,
-            LineString.class_from_string(
-                '<LineString xmlns="http://www.opengis.net/kml/2.2">'
-                "<extrude>1</extrude>"
-                "<tessellate>1</tessellate>"
-                "<coordinates>"
-                "-122.364383,37.824664,0 -122.364152,37.824322,0"
-                "</coordinates>"
-                "</LineString>",
-            ),
+        linestring = LineString.class_from_string(
+            '<LineString xmlns="http://www.opengis.net/kml/2.2">'
+            "<extrude>1</extrude>"
+            "<tessellate>1</tessellate>"
+            "<coordinates>"
+            "-122.364383,37.824664,0 -122.364152,37.824322,0"
+            "</coordinates>"
+            "</LineString>",
         )
 
         assert linestring.geometry == geo.LineString(
@@ -99,30 +95,24 @@ class TestLineString(StdLibrary):
 
     def test_empty_from_string(self) -> None:
         """Test the from_string method with an empty LineString."""
-        linestring = cast(
-            LineString,
-            LineString.class_from_string(
-                '<LineString xmlns="http://www.opengis.net/kml/2.2">'
-                "<extrude>1</extrude>"
-                "<tessellate>1</tessellate>"
-                "<coordinates>"
-                "</coordinates>"
-                "</LineString>",
-            ),
+        linestring = LineString.class_from_string(
+            '<LineString xmlns="http://www.opengis.net/kml/2.2">'
+            "<extrude>1</extrude>"
+            "<tessellate>1</tessellate>"
+            "<coordinates>"
+            "</coordinates>"
+            "</LineString>",
         )
 
         assert linestring.geometry.is_empty
 
     def test_no_coordinates_from_string(self) -> None:
         """Test the from_string method with no coordinates."""
-        linestring = cast(
-            LineString,
-            LineString.class_from_string(
-                '<LineString xmlns="http://www.opengis.net/kml/2.2">'
-                "<extrude>1</extrude>"
-                "<tessellate>1</tessellate>"
-                "</LineString>",
-            ),
+        linestring = LineString.class_from_string(
+            '<LineString xmlns="http://www.opengis.net/kml/2.2">'
+            "<extrude>1</extrude>"
+            "<tessellate>1</tessellate>"
+            "</LineString>",
         )
 
         assert linestring.geometry.is_empty
